+++ src/base_dumper/common/dump_format.py
@@ -1,0 +1,9 @@
+from enum import Enum
+
+
+class DumpFormat(Enum):
+    """Enum for dump type format."""
+
+    RAW = 0
+    CSV = 1
+    S3 = 2